requirements.txt
Django>=5.0
djangorestframework
django-filter
drf-spectacular
psycopg[binary]
python-dotenv
openpyxl>=3.1.0
"django-redis>=6.0.0"  
"redis>=7.0.0" 
